requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-MySQLdb==0.2.0
greenlet==1.0.0
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.20.1
pathspec==0.8.1
protobuf==3.15.6
regex==2021.3.17
six==1.15.0
SQLAlchemy==1.4.2
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.3
Werkzeug==1.0.1
